REF: share GroupBy.count
